+++ application/next-frontend/pages/index.tsx
@@ -1,12 +1,11 @@
-/* import { Landing } from '@/Landing' */
+import { Landing } from '@/Landing'
 import { Header } from '@/Landing/scenarios/Header'
 
 export default function Home() {
     return (
         <div className="h-full">
             <Header />
-            <h1 className="mt-12 text-center font-workSans text-3xl">COMING SOON</h1>
-            {/* <Landing /> */}
+            <Landing />
         </div>
     )
 }
